+++ src/main/java/pokemones/tipoPlanta.java
@@ -6,7 +6,7 @@ import movimientos.Movimientos;
 public class tipoPlanta extends Mounstritos {
 
     private String[] debilidades;
-    public tipoPlanta(String name, double HP, double maxHP, ListaDE<Movimientos> ataques, int defense, int ataque, String sprite, boolean vivo) {
+    public tipoPlanta(String name, double HP, double maxHP, ListaDE<Movimientos> ataques, float defense, int ataque, String sprite, boolean vivo) {
         super(name, HP, maxHP, ataques, defense, ataque, sprite, vivo,new String[]{"Fuego", "Aire"});
 
     }
